+++ ModConfig.cs
@@ -220,6 +220,8 @@ sealed class ModConfig(ConfigFile cfg)
         public ConfigEntry<Emotes> StackInventoryIntoContainersEmote { get; } = cfg.Bind(section, nameof(StackInventoryIntoContainersEmote), DisabledEmote,
             new ConfigDescription($"Emote to stack inventory into containers. {DisabledEmote} to disable this feature, {AnyEmote} to use any emote as trigger", new AcceptableEnum<Emotes>([DisabledEmote, AnyEmote, .. Enum.GetValues(typeof(Emotes)).Cast<Emotes>()])));
 
+        public ConfigEntry<bool> CanSacrificeMegingjord { get; } = cfg.Bind(section, nameof(CanSacrificeMegingjord), false,
+            "If true, players can permanently unlock increased carrying weight by sacrificing a megingjord in an obliterator");
         public ConfigEntry<bool> CanSacrificeCryptKey { get; } = cfg.Bind(section, nameof(CanSacrificeCryptKey), false,
             "If true, players can permanently unlock the ability to open sunken crypt doors by sacrificing a crypt key in an obliterator");
         public ConfigEntry<bool> CanSacrificeWishbone { get; } = cfg.Bind(section, nameof(CanSacrificeWishbone), false,